Update travis script
This adds the ability have Travis cache bundler (when supported) to help
speed up builds. It also updates the bundler version to the latest;
sadly Travis-CI needs this explicit gem install line otherwise it'll use
a really old version of bundler.

This also includes more Ruby versions to test. We test the "latest" of
each minor version of Ruby 2.x to ensure compatibility while keeping the
build matrix small. We need to specific the full version because it
doesn't appear that Travis-CI is updating their build alias names (i.e.
Ruby 2.2 is only Ruby 2.2.0). To prepare for the next Ruby release this
also includes the `ruby-head` version.  However, as it's a future
release we allow it to fail while still considering the build as
passing.

Lastly, while all we do currently for setup is run bundler this could
change in the future. To be consistent in setup we now use the
`bin/setup` script for the CI install.